use MutableList instead of List and update score
