--- app/src/main/java/com/cs125/anappleaday/data/record/RelaxationData.kt
@@ -1,9 +1,0 @@
-package com.cs125.anappleaday.data.record
-
-data class RelaxationData(
-    val id: String,
-
-    val scoreData: ScoreData,
-
-    val relaxationRecords: List<RelaxationRecord>
-)
