--- app/src/main/java/com/cs125/anappleaday/data/record/ScoreData.kt
@@ -1,7 +1,0 @@
-package com.cs125.anappleaday.data.record
-
-data class ScoreData(
-    val avgScore: Double,
-
-    val scores: List<ScoreDay>
-)
