--- app/src/main/java/com/cs125/anappleaday/data/record/SleepData.kt
@@ -1,11 +1,0 @@
-package com.cs125.anappleaday.data.record
-
-import android.health.connect.datatypes.SleepSessionRecord
-
-data class SleepData(
-    val id: String,
-
-    val scoreData: ScoreData,
-
-    val sleepSessionRecords: List<SleepSessionRecord>
-)
