+++ app/src/main/java/com/cs125/anappleaday/data/record/models/ActivityData.kt
@@ -1,11 +1,11 @@
-package com.cs125.anappleaday.data.record
+package com.cs125.anappleaday.data.record.models
 
 import android.health.connect.datatypes.ExerciseSessionRecord
 
 data class ActivityData(
     val id: String,
 
-    val scoreData: ScoreData,
+    val scores: MutableList<ScoreDay>,
 
     val activities: List<ExerciseSessionRecord>,
 )
