+++ app/src/main/java/com/cs125/anappleaday/data/record/models/MedicalRecord.kt
@@ -1,0 +1,11 @@
+package com.cs125.anappleaday.data.record.models
+
+data class MedicalRecord(
+    val id: String,
+
+    val allergies: MutableList<String>,
+
+    val foodRestrictions: MutableList<String>,
+
+    val disorders: MutableList<String>
+)
