+++ app/src/main/java/com/cs125/anappleaday/data/record/models/RelaxationData.kt
@@ -1,0 +1,9 @@
+package com.cs125.anappleaday.data.record.models
+
+data class RelaxationData(
+    val id: String,
+
+    val scores: MutableList<ScoreDay>,
+
+    val relaxationRecords: MutableList<RelaxationRecord>
+)
