+++ app/src/main/java/com/cs125/anappleaday/data/record/models/RelaxationRecord.kt
@@ -1,4 +1,4 @@
-package com.cs125.anappleaday.data.record
+package com.cs125.anappleaday.data.record.models
 
 import com.cs125.anappleaday.data.enumTypes.StressLevel
 import java.util.Date
